# avoid intermediate array for lz4/zlib