When decompressing, we do something like

`@view(uncomp_data[fufilled+1:fufilled+uncompbytes]) = <output of decompression algo>`

for LZ4, this is actually making a temporary array `out_buffer` which is then copied into `uncomp_data`.

```julia
function lz4_decompress(
        input::Union{Vector{UInt8},Base.CodeUnits{UInt8}},
        expected_size::Integer=length(input) * 2
    )
    out_buffer = Vector{UInt8}(undef, expected_size)
    out_size = LZ4_decompress_safe(pointer(input), pointer(out_buffer), length(input), expected_size)
    resize!(out_buffer, out_size)
end
```

By feeding `LZ4_decompress_safe` a pointer to `uncomp_data`, we can go faster! :rocket: 
And the same exercise was repeated for ZLIB to eliminate the intermediate storage. I copied almost exactly what ROOT has.

## LZ4

#### before

```julia
julia> @btime sum(tf.nMuon)
  600.881 ms (2650 allocations: 870.01 MiB)
0x0000000008e67ad8
```

#### after

```julia
julia> @btime sum(tf.nMuon)
  472.534 ms (2339 allocations: 580.06 MiB)
0x0000000008e67ad8
```


## ZLIB

#### before

```julia
julia> @btime sum(tf.nMuon)
  1.326 s (15170 allocations: 848.43 MiB)
0x0000000008e67ad8


julia> @time sum.(tf.Muon_pt)
  8.802347 seconds (22.05 k allocations: 4.399 GiB, 6.27% gc time)
```

#### after

```julia
julia> @btime sum(tf.nMuon)
  1.107 s (12162 allocations: 544.53 MiB)
0x0000000008e67ad8

julia> @time sum.(tf.Muon_pt)
  7.474018 seconds (18.92 k allocations: 3.216 GiB, 5.11% gc time)

```

......so ~15% speedups for zlib and more for lz4.