@aminnj 2021-10-05
For future reference, this is how ROOT does zlib for example, but the others are in there as well:

https://github.com/root-project/root/blob/87a998d48803bc207288d90038e60ff148827664/core/zip/src/RZip.cxx#L392