@Moelf 2021-10-05
do we want to wrap this whole if-else logic into a function? since technically there's another place we use this:
https://github.com/tamasgal/UnROOT.jl/blob/35d00838067d68e74ede4ebddfb298c4eac4f813/src/streamers.jl#L85-L87