@aminnj 2021-10-05
in principle, yes, we can consolidate them, though I don't know if it's worth it because the streamers are only unpacked once and they're small. Maybe once we're happy with the inplace versions (and perhaps also have done if for the remaining two algos as well)