@aminnj 2021-10-05
And for reference, the updated zlib decompression function works with [zlib-ng](https://github.com/zlib-ng/zlib-ng) still. Eg

```bash
cp ~/julia-1.7.0-beta4/lib/julia/libz.so.1.2.11 bck.so # backup
# compile zlib-ng
cp ~/zlib-ng/libz.so.1.2.11.zlib-ng ~/julia-1.7.0-beta4/lib/julia/libz.so.1.2.11
```

and gives another 10% improvement in decompression speed itself. So when that gets incorporated, we get another speedboost.